test: Add unit test for onchain delegates
